Fixing float / double mismatches.
--HG--
branch : week-of-code

+++ src/enzo/Grid_GalaxySimulationInitializeGrid.C
@@ -56,7 +56,7 @@ void rot_to_disk(FLOAT xpos, FLOAT ypos, FLOAT zpos, FLOAT &xrot, FLOAT &yrot, F
 /* Internal Routines for Disk Potential Setup */
 float HaloGasDensity(FLOAT);
 float HaloGasTemperature(FLOAT);
-double DiskPotentialCircularVelocity(FLOAT cellwidth,FLOAT z,FLOAT density,FLOAT &temperature);
+float DiskPotentialCircularVelocity(FLOAT cellwidth,FLOAT z,FLOAT density,FLOAT &temperature);
 double trapzd(double (func)(), double a, double b, int n);
 double qromb(double (*func)(double), double a, double b);
 void polint(double xa[],double ya[],int n,double x,double *y,double *dy);
